# src/shs.Domain/Enums/ConsignmentStatusType.cs
namespace shs.Api.Domain.Enums;

public enum ConsignmentStatusType
{
    Evaluated = 1,
    AwaitingAcceptance = 2,
    ToSell = 3,
    Sold = 4,
    Returned = 5
}

public enum ProductSizeType
{
    Small = 1,
    Medium = 2,
    Large = 3
}